update: (#447) studyroom port 병합
