--- dms-core/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/AvailableTimePort.kt
@@ -1,5 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-interface AvailableTimePort :
-    QueryAvailableTimePort,
-    CommandAvailableTimePort
