--- dms-core/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/CommandAvailableTimePort.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-import team.aliens.dms.domain.studyroom.model.AvailableTime
-
-interface CommandAvailableTimePort {
-
-    fun saveAvailableTime(availableTime: AvailableTime): AvailableTime
-}
