--- dms-core/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/CommandSeatTypePort.kt
@@ -1,10 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-import team.aliens.dms.domain.studyroom.model.SeatType
-
-interface CommandSeatTypePort {
-
-    fun saveSeatType(seatType: SeatType): SeatType
-
-    fun deleteSeatType(seatType: SeatType)
-}
