--- dms-core/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/QueryAvailableTimePort.kt
@@ -1,9 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-import team.aliens.dms.domain.studyroom.model.AvailableTime
-import java.util.UUID
-
-interface QueryAvailableTimePort {
-
-    fun queryAvailableTimeBySchoolId(schoolId: UUID): AvailableTime?
-}
