--- dms-core/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/QuerySeatTypePort.kt
@@ -1,15 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-import team.aliens.dms.domain.studyroom.model.SeatType
-import java.util.UUID
-
-interface QuerySeatTypePort {
-
-    fun queryAllSeatTypeBySchoolId(schoolId: UUID): List<SeatType>
-
-    fun queryAllSeatTypeByStudyRoomId(studyRoomId: UUID): List<SeatType>
-
-    fun existsSeatTypeByName(name: String): Boolean
-
-    fun querySeatTypeById(seatTypeId: UUID): SeatType?
-}
