--- dms-core/src/main/kotlin/team/aliens/dms/domain/studyroom/spi/SeatTypePort.kt
@@ -1,5 +1,0 @@
-package team.aliens.dms.domain.studyroom.spi
-
-interface SeatTypePort :
-    QuerySeatTypePort,
-    CommandSeatTypePort
